pyproject.toml
[project]
name = "cognify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-google-genai>=2.1.12",
    "langgraph>=0.6.7",
    "python-dotenv>=1.1.1",
]
